REFACTOR: MainForm and MainDataForm components completely replace old CurpComponent, CurpByDataComponent and RfcFormComponent and RfcDataFormComponent.
