+++ src/app/core/main-form/main-data-form/main-data-form.component.scss
@@ -1,0 +1,4 @@
+// :host {
+//     display: block;
+//     height: 26rem;
+// }
