+++ src/app/core/main-form/main-data-form/rfc-data-form/rfc-data-form.component.scss
@@ -1,0 +1,3 @@
+p-datepicker, p-autocomplete {
+    width: 100%;
+}
