+++ src/app/core/main-form/main-form.component.scss
@@ -13,4 +13,5 @@
     justify-content: space-between;
     width: 100%;
     height: 100%;
+    position: relative;
 }
