--- src/app/core/rfc-form/query-cp-form/query-cp-form.component.scss
@@ -1,3 +1,0 @@
-:host {
-    height: 100%;
-}
