--- src/app/core/rfc-form/rfc-data-form/rfc-data-form.component.scss
@@ -1,3 +1,0 @@
-p-datepicker {
-    width: 100%;
-}
