--- src/app/core/rfc-form/rfc-form.component.scss
@@ -1,4 +1,0 @@
-:host {
-    height: 100%;
-    width: 100%;
-}
