--- src/app/core/rfc-form/rfc-form.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { RfcFormService } from './rfc-form.service';
-
-describe('RfcFormService', () => {
-  let service: RfcFormService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(RfcFormService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
